Move BeyondFrame component to diff file
